api/requirements.txt
fastapi==0.115.0
fastapi-cli==0.0.5
pydantic==2.9.2
psycopg2-binary==2.9.9
Werkzeug==3.0.4
